+++ tier1/KeyValues.cpp
@@ -1830,7 +1830,7 @@ void KeyValues::CopyKeyValue( const KeyValues& src, size_t tmpBufferSizeB, char*
 	case TYPE_INT:
 		{
 			m_iValue = src.m_iValue;
-			Q_snprintf( tmpBuffer, tmpBufferSizeB, "%d", m_iValue );
+			Q_snprintf( tmpBuffer, static_cast<int>(tmpBufferSizeB), "%d", m_iValue );
 			int len = Q_strlen(tmpBuffer) + 1;
 			m_sValue = new char[len];
 			Q_strncpy( m_sValue, tmpBuffer, len  );
@@ -1839,7 +1839,7 @@ void KeyValues::CopyKeyValue( const KeyValues& src, size_t tmpBufferSizeB, char*
 	case TYPE_FLOAT:
 		{
 			m_flValue = src.m_flValue;
-			Q_snprintf( tmpBuffer, tmpBufferSizeB, "%f", m_flValue );
+			Q_snprintf( tmpBuffer, static_cast< int >( tmpBufferSizeB ), "%f", m_flValue );
 			int len = Q_strlen(tmpBuffer) + 1;
 			m_sValue = new char[len];
 			Q_strncpy( m_sValue, tmpBuffer, len );
