+++ tier1/utlstring.cpp
@@ -52,7 +52,7 @@ void CUtlString::SetDirect( const char *pValue, int nChars )
 			return; // Do nothing. Realloc in AllocMemory might move pValue's location resulting in a bad memcpy.
 		}
 
-		Assert( nChars <= Min<int>( strnlen(pValue, nChars) + 1, nChars ) );
+		Assert( nChars <= Min<int>( static_cast< int >( strnlen(pValue, nChars) + 1 ), nChars ) );
 		AllocMemory( nChars );
 		Q_memcpy( m_pString, pValue, nChars );
 	}
@@ -624,7 +624,7 @@ CUtlString CUtlString::Replace( const char *pszFrom, const char *pszTo ) const
 	size_t srcLength = Length();
 	CUtlString strDest;
 	size_t destLength = srcLength + nAllocOffset;
-	strDest.SetLength( destLength );
+	strDest.SetLength( static_cast< int >( destLength ) );
 
 	// find and replace the search string
 	pos = pFirstFound;
@@ -648,7 +648,7 @@ CUtlString CUtlString::Replace( const char *pszFrom, const char *pszTo ) const
 	Assert( destLength - nDestOffset == srcLength - nSrcOffset );
 	if ( destLength - nDestOffset > 0 )
 	{
-		V_strncpy( strDest.GetForModify() + nDestOffset, String() + nSrcOffset, destLength - nDestOffset + 1 );
+		V_strncpy( strDest.GetForModify() + nDestOffset, String() + nSrcOffset, static_cast< int >( destLength ) - nDestOffset + 1 );
 	}
 
 	return strDest;
